* commit 'cd79e6d3f3059a2d513c65cb7fbc4295933e1f68':
  Add event management to file system unlink file process
  Add event management to download/upload process
  risolto #4325 - Rimuovere vecchio template docebo 405
  fixed: #8249 La treeview si basava su un url che puntava alla index di appCore da appLms. Corretto aggiungendo un link relativo.
  added a missing affected_rows() function
  #4229 ripristinato customfield di tipo testo perchè sistemato add edit e delete in calsse
  #4229 In progress - Sistemato add mod e delete di classe textfield dei customfield
  Risolto #7653 Aggiunto script sql per visualizzare anche add mod e del in gestione amministratore semplice
  Ripristinati filtro ricerca e paginazione.
  fixed: #8035 Ora premendo sul tasto fullscreen di tinymce la finestra non viene visualizzata sotto l'header di forma
  Commentata la parte che fa l'update della password. Rimane comunque funzionante per chi avesse già aggiornato db e password.
  risolto #7653 Profilo amministratore - visualizzazione saltata - wrong visualization under Administrator Profile
  #3962 cleaning CMS code references
  File sql per aggiornare la colonna pass di core_user. Per SHA256 sono necessari 60 caratteri.
  Fix course forum thread
  Fix course forum thread
  Cartella mancante per upload immagini caricate da utenti. Cambiato percorso da relativo ad assoluto (utilizzando costante). Causava un problema ad TinyMCE
  Il sistema utilizzato per la generazione delle password non sarebbe compatibile con PHP < 5.5. Includendo questa libreria viene risolto il problema. Verrà, in futuro, rimosso da "addons/"
  Fixed IE11 issues: 1) Scorm required forcing IE9 compatibility but this caused other problems; now scorm works fine under IE11 without 2) Some button did not work
  Ora l'algoritmo  di hash per le password di default di forma è SHA-256. E' presente un meccanismo che aggiorna automaticamente le password in md5 ogni qual volta una delle seguenti azioni è eseguita - api (generate token) - login - cambio pass da utente - creazione utente da admin - modifica password da admin - richiesta modifica password da admin - ldap

+++ html/lib/lib.upload.php
@@ -69,7 +69,7 @@ function sl_fopen( $filename, $mode ) {
 }
 
 function sl_upload( $srcFile, $dstFile, $file_ext) {
-	$uploadType = Get::cfg('uploadType');
+	$uploadType = Get::cfg('uploadType', null);
 
 	// check if the mime type is allowed by the whitelist
 	// if the whitelist is empty all types are accepted
@@ -109,9 +109,14 @@ function sl_upload( $srcFile, $dstFile, $file_ext) {
 		return sl_upload_ftp( $srcFile, $dstFile );
 	} elseif( $uploadType == "cgi" ) {
 		return sl_upload_cgi( $srcFile, $dstFile );
-	} else {
+	} elseif( $uploadType == "fs" || $uploadType == null ) {
 		return sl_upload_fs( $srcFile, $dstFile );
-	}
+	} else {
+        $event = new \appCore\Events\Core\FileSystem\UploadEvent($srcFile, $dstFile);
+        \appCore\Events\DispatcherManager::dispatch(\appCore\Events\Core\FileSystem\UploadEvent::EVENT_NAME, $event);
+
+        return $event->getResult();
+    }
 }
 
 function sl_touch( $filename, $time ) {
@@ -172,8 +177,8 @@ function sl_copy( $srcFile, $dstFile ) {
 
 function sl_upload_fs( $srcFile, $dstFile ) {
 
-	$re = move_uploaded_file($srcFile, $GLOBALS['where_files_relative'].$dstFile);
-	if(!$re) die("Error on move_uploaded_file from: $srcFile to $dstFile");
+	$re = move_uploaded_file($srcFile, _files_.$dstFile);
+	if(!$re) die("Error on move_uploaded_file from: $srcFile to ".$dstFile);
 	return $re;
 }
 
@@ -331,9 +336,19 @@ function sl_upload_cgi( $srcFile, $dstFile ) {
  **/
 
  function sl_unlink( $path ) {
+     $uploadType = Get::cfg('uploadType', null);
 
-	 if( !file_exists($GLOBALS['where_files_relative'].$path) ) return true;
-	 return @unlink($GLOBALS['where_files_relative'].$path);
+     if( $uploadType == "fs" || $uploadType == null ) {
+         if( !file_exists($GLOBALS['where_files_relative'].$path) ) return true;
+         return @unlink($GLOBALS['where_files_relative'].$path);
+     } else {
+         $event = new \appCore\Events\Core\FileSystem\UnlinkEvent($path);
+         \appCore\Events\DispatcherManager::dispatch(\appCore\Events\Core\FileSystem\UnlinkEvent::EVENT_NAME, $event);
+
+         return $event->getResult();
+     }
+
+
  }
 
 ?>
